[boost] Use C++11 stdlib classes instead of boost.
