+++ src/xpcc/processing/rtos/mutex.hpp
@@ -34,7 +34,7 @@
 #include <xpcc/architecture/utils.hpp>
 
 #ifdef XPCC__OS_HOSTED
-#	include "boost/mutex.hpp"
+#	include "stdlib/mutex.hpp"
 #elif defined(XPCC__CPU_CORTEX_M3) || defined(XPCC__CPU_CORTEX_M4)
 #	include "freertos/mutex.hpp"
 #endif
