+++ src/xpcc/processing/rtos/stdlib/queue_impl.hpp
@@ -32,6 +32,8 @@
 #	error "Don't use this file directly, use 'queue.hpp' instead!"
 #endif
 
+#include <chrono>
+
 template <typename T>
 xpcc::rtos::Queue<T>::Queue(uint32_t length) :
 	maxSize(length)
@@ -47,7 +49,7 @@ template <typename T>
 std::size_t
 xpcc::rtos::Queue<T>::getSize() const
 {
-	boost::lock_guard<boost::timed_mutex> lock(mutex);
+	std::lock_guard<std::timed_mutex> lock(mutex);
 	deque.size();
 }
 
@@ -55,7 +57,7 @@ template <typename T>
 bool
 xpcc::rtos::Queue<T>::append(const T& item, uint32_t timeout = -1)
 {
-	if (!mutex.timed_lock(boost::posix_time::milliseconds(timeout)) ||
+	if (!mutex.try_lock_for(std::chrono::milliseconds(timeout)) ||
 			deque.size() >= maxSize) {
 		return false;
 	}
@@ -68,7 +70,7 @@ template <typename T>
 bool
 xpcc::rtos::Queue<T>::prepend(const T& item, uint32_t timeout = -1)
 {
-	if (!mutex.timed_lock(boost::posix_time::milliseconds(timeout))) {
+	if (!mutex.try_lock_for(std::chrono::milliseconds(timeout))) {
 		return false;
 	}
 	
@@ -88,7 +90,7 @@ template <typename T>
 bool
 xpcc::rtos::Queue<T>::peek(T& item, uint32_t timeout = -1) const
 {
-	if (!mutex.timed_lock(boost::posix_time::milliseconds(timeout))) {
+	if (!mutex.try_lock_for(std::chrono::milliseconds(timeout))) {
 		return false;
 	}
 	
@@ -107,7 +109,7 @@ template <typename T>
 bool
 xpcc::rtos::Queue<T>::get(T& item, uint32_t timeout = -1)
 {
-	if (!mutex.timed_lock(boost::posix_time::milliseconds(timeout))) {
+	if (!mutex.try_lock_for(std::chrono::milliseconds(timeout))) {
 		return false;
 	}
