+++ src/xpcc/processing/rtos/stdlib/semaphore.hpp
@@ -35,8 +35,8 @@
 #	error "Don't include this file directly, use <xpcc/processing/rtos/semaphore.hpp>"
 #endif
 
-#include <boost/thread/mutex.hpp>    
-#include <boost/thread/condition_variable.hpp>
+#include <mutex>
+#include <condition_variable>
 
 namespace xpcc
 {
@@ -125,10 +125,10 @@ namespace xpcc
 			// 
 			// Any code that reads or writes the count_ data must hold a lock
 			// on the mutex.
-			mutable boost::mutex mutex;
+			mutable std::mutex mutex;
 			
 			// Code that increments count_ must notify the condition variable.
-			mutable boost::condition_variable condition;
+			mutable std::condition_variable condition;
 
 		};
