+++ src/xpcc/processing/rtos/thread.hpp
@@ -34,7 +34,7 @@
 #include <xpcc/architecture/utils.hpp>
 
 #ifdef XPCC__OS_HOSTED
-#	include "boost/thread.hpp"
+#	include "stdlib/thread.hpp"
 #elif defined(XPCC__CPU_CORTEX_M3) || defined(XPCC__CPU_CORTEX_M4)
 #	include "freertos/thread.hpp"
 #endif
